Brian begin slide

+++ src/main/java/com/project/sorryapp/GatewayTile.java
@@ -5,7 +5,7 @@ import javafx.scene.paint.Color;
 public class GatewayTile extends Tile{
     private Tile gatewayNext_;
 
-    GatewayTile(int x, int y, double length, Color color){
+    GatewayTile(double x, double y, double length, Color color){
         super(x,y,length,color);
         this.set_moveBehavior(new GatewayMove());
     }
